tests: factor lexing operations
