+++ test/test_utils/CMakeLists.txt
@@ -1,4 +1,7 @@
 target_sources(${PROJECT_TEST_NAME}
+    PRIVATE
+        "${CMAKE_CURRENT_LIST_DIR}/lexing.cc"
     PUBLIC
+        "${CMAKE_CURRENT_LIST_DIR}/lexing.h"
         "${CMAKE_CURRENT_LIST_DIR}/utils.h"
     )
